gcp/pipelines/transformation/requirements.txt
flask==3.*
google-cloud-storage==2.*
gunicorn==21.*